requirements.txt
fastapi
uvicorn
openai
python-dotenv
flatlib
pyswisseph
pydantic
requests